Unity Practice Panels
These should work for the basic. They are just cubes of the correct dimensions, but they have the correct centers.
In STR_UnityPanels I set the bottom-center of each panel to 0, 0, 0.
In STR_UnityPanelsCornerPivot I set one of the bottom corners of each panel to 0, 0, 0.
Changing the pivot point from Center to Pivot alters the pivot point of the objects form the center of the objects to their pivot point. See the picture titled PivotSetting for more info and/or check out this video: https://www.youtube.com/watch?v=MpvwdOCvC9k